refactor: config 패키지 통합 및 초기 데이터 주입 패키지 변경
